Manage the Python development environment using Poetry
Poetry is a tool for dependency management and packaging in Python. It allows you to declare the libraries your project depends on, and it will manage (install/update) them for you. Poetry offers a lock file to ensure repeatable installs and can build your project for distribution. This update will speed up the installation process of the Python development environment used in this course.